Refactor MapBuilderContext to reduce code duplication. (#1059)
